Refactorise el botón de inicio de sesión y los componentes de usuario para utilizar la utilidad getReturnUrl para la gestión de URL
