+++ src/components/menu/button-login.tsx
@@ -1,15 +1,10 @@
 "use client"
 import Link from "next/link"
 import * as Styles from "@/styles/menu-home-style"
-import { usePathname } from 'next/navigation';
+import { getReturnUrl } from "@/utils/get-return-url";
 
 export default function ButtonLogin() {
-    const pathname = usePathname();
-    const base = process.env.NODE_ENV === 'development' 
-        ? 'http://localhost:3000' 
-        : process.env.NEXT_PUBLIC_URL_BASE;
-
-    const url = `${base}/account/login?return_url=${encodeURIComponent(base + pathname)}`;
+    const { url } = getReturnUrl();
 
     return (
         <Link href={url}>
